requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
psycopg2-binary==2.9.1
pandas==1.3.3
numpy==1.21.2
gunicorn==20.1.0
python-dotenv==0.19.0
